Added new GSetting schema and keys to indepentently control scaling factor
Until now, we were directly modifying the value of the GSetting key
text-scaling-factor, from the org.gnome.desktop.interface schema, to
manipulate the effective DPI of the screen when on composite mode, which
caused "collisions" with the 'Larger Text' accessibility feature, as it
would be marked as enabled when this value is larger than 1.0.

To avoid this problem, and also to be more flexible so that we can
define our own Composite-related scaling factors if needed, we define a
new GSetting schema com.endlessm.CompositeMode have this daemon changing
only the value of a 'text-scaling-factor' key there, leaving the one from
org.gnome.desktop.interface untouched. This way, it will be now the
responsibility of gnome-settings-daemon to read that new key if
available and use its value to set the effective DPI value, which will
be used from any applicationreading the GtkSettings 'gtk-xft-dpi' property.

Also, as a consecuence, we no longer need the persistance features added
in previous commits, so I'm removing ecm_settings_persist() too.

[endlessm/eos-shell#6197]