Removed the initialized state and introduced new schemas for default values
Also refactored a bit the code and merged everything back into ecm_daemon.c,
as now code is simpler and we don't really need that much separation (and
it's also nice that we can reuse one GSettings object).

Last, bumped the default scaling factor for the browser to 1.2 so that it
will result in a 130% increase when applied on top of text-scaling-factor,
as chromium won't do any scaling for smaller factors (crbug.com/484400).

[endlessm/eos-shell#6197]