Do not set the scaling settings if they are equal to the current ones
Somehow always setting the scaling values is resulting in the settings
change event getting called constantly (not in every boot though) with
the same values.
I am not sure of the root cause for this taking into account that this
issue is very tricky and that it's been investigated for a while now,
I am proposing this patch that skips changing the gsettings if the new
values are already set.

https://phabricator.endlessm.com/T18078